refactor: using cms for data instead of gatsby-source-filsystem
Production builds went from 5 to 22 minutes after adding chairs.
