+++ src/components/X-Chair/generateInitialState.js
@@ -5,12 +5,84 @@ function generateColorCheckboxes(num) {
   temp[0] = true;
   return temp;
 }
-const GenerateInitialState = (
-  colorCB,
-  activeColor,
-  chairVariants,
-  chairData
-) => {
+function generateColorData(mainColors, extra1, extra2) {
+  // combines gallery acrossed each fabric and creates color swatchs for color feature component
+  const colorData = {};
+  const mainColorSwatchs = [];
+  const extraColorSwatchs = [];
+  const extraColorSwatchs2 = [];
+  let extraColors = null;
+  let indexCount = 0;
+  if (extra1.length !== 0 && extra2.length !== 0) {
+    mainColors.forEach((element) => {
+      colorData[element.colorTitle] = {
+        default: element.default,
+        headrest: element.withHeadrestImages,
+      };
+      mainColorSwatchs.push({
+        title: element.colorTitle,
+        img: element.colorSwatch.gatsbyImageData,
+        alt: element.colorSwatch.alt,
+        index: indexCount ? indexCount + 1 : indexCount,
+      });
+      indexCount += 1;
+    });
+    extra1.forEach((element) => {
+      colorData[element.colorTitle] = {
+        default: element.default,
+        headrest: element.withHeadrestImages,
+      };
+      extraColorSwatchs.push({
+        title: element.colorTitle,
+        img: element.colorSwatch.gatsbyImageData,
+        alt: element.colorSwatch.alt,
+        index: indexCount ? indexCount + 1 : indexCount,
+      });
+      indexCount += 1;
+    });
+    extra2.forEach((element) => {
+      colorData[element.colorTitle] = {
+        default: element.default,
+        headrest: element.withHeadrestImages,
+      };
+      extraColorSwatchs2.push({
+        title: element.colorTitle,
+        img: element.colorSwatch.gatsbyImageData,
+        alt: element.colorSwatch.alt,
+        index: indexCount ? indexCount + 1 : indexCount,
+      });
+      indexCount += 1;
+    });
+    extraColors = {
+      "Premium Leather": extraColorSwatchs,
+      Brisa: extraColorSwatchs2,
+    };
+  } else {
+    mainColors.forEach((element, index) => {
+      colorData[element.colorTitle] = {
+        default: element.default,
+        headrest: element.withHeadrestImages,
+      };
+      mainColorSwatchs.push({
+        title: element.colorTitle,
+        img: element.colorSwatch.gatsbyImageData,
+        alt: element.colorSwatch.alt,
+        index,
+      });
+    });
+  }
+  return {
+    allColorData: colorData,
+    mainSwatches: mainColorSwatchs,
+    extraColors,
+  };
+}
+const GenerateInitialState = (chairData) => {
+  const colorData = generateColorData(
+    chairData.colors,
+    chairData.premiumLeather,
+    chairData.brisa
+  );
   const initialState = {
     activeColor: chairData.colors[0].colorTitle,
     activeChairIndex: 0,
@@ -31,7 +103,7 @@ const GenerateInitialState = (
     activeWheelIndex: 0,
     activeModelIndex: 0,
   };
-  return initialState;
+  return { initialState, colorData };
 };
 
 export default GenerateInitialState;
